refactor(components): change component file names
