+++ src/components/Content.tsx
@@ -1,8 +1,8 @@
 import { FC, ReactNode } from 'react';
-import { Status } from '../../constants/enums';
-import { Loading } from '../../interfaces/Loading';
-import Alert from '../Alert';
-import Progress from '../Progress';
+import { Status } from '../constants/enums';
+import { Loading } from '../interfaces/Loading';
+import Alert from './Alert';
+import Progress from './Progress';
 
 interface ContentProps extends Loading {
   children: ReactNode;
